refactor(project): re-structure the entire project and add poetry.lock for dep. management
poetry

BREAKING CHANGE: file names and paths need to be changed in modules
